Set the root namespace to a used namespace.
The problem is that the R.cs-generation behavior changed so that it's always
emitted into the namespace defined by the <RootNamespace/> element.
Unfortunately, <RootNamespace/> was set to "MineSweeper" and never used
anywhere.  Result: references to 'R' generated compilation errors.

Oops.

Changing the <RootNamespace/> to something that is used by the code results in
R being in scope, and no compilation errors.