[Support4] Fix Color-related breakage.
Interestingly, the Color changes found a bug. This isn't valid:

	new ColorDrawable(Resource.Drawable.background)

It's not valid because ColorDrawable wants a color.
Resource.Drawable.background isn't a color; it's a color reference.
(Compare the value of Resource.Drawable.background -- 0x7f020007 -- to
the desired value -- #ff0000.)

To get the actual corresponding color you need to use
Resources.GetColor(int).