[AdMob] Update to bind AdMob v6.0.0.
We had to use `argsType` to rename the generated EventArgs class for
each event handler, because the MediationBannerListener and
MediationInterstitialListener types had numerious methods with the
same name, and (for clarity) the default EventHandler naming algorithm
considers only the method name, not the interface name.

We also hide the internal types, as (1) they were generating
compilation errors (e.g. a 'c.c' type), and (2) they're internal
(even if they're "public" in the .jar).