Add OpenTK 1.0 samples.
We'll be merging the monodroid-4.1-samples branch with master, and we
need to keep two sets of OpenTK samples around: one for testing the
original OpenTK.dll (v0.9) and one testing the new OpenTK-1.0.dll.

Copy the OpenTK-1.0-targeting samples into new directories so that
they'll survive the merge unchanged.