Fragment Fix
To change the content of the view by selecting a different tab, it is necessary to remove the old content and then add the new because if you don't remove first, both will be shown. As long as both view have the same content noone gets to know, but if both Tabs have a different content both will be shown after a selection change.