Fix: Orientation Change brakes tabs
If an Orientation Change is performed, the Tab changes back to the first one even if a differnt one was selected. Further more, the old Tab which was displayed before the change will be displayed behind the new one so that there are two Tabs which are displayed at the same time.