[KitKat] Added Toast Response to sensor start
When using a device that did have a step counter built in, the user
would not be able to tell that the button did anything because there
was no visual response to hitting it. This also solves test failures as
it gives a message to check for on these devices.