Fix for XA 5.1 and remove async code
LinedEditText layout class needed to be registered in order to work with XA 5. Removed the async loading and saving of notes as it was unnecessary and was causing nullptr exceptions when loading note descriptions. Also modified the getting and putting of note Id longs when navigating between activities. It was throwing a conversion error, but did not crash the app.