Update minSdkVersion for projects using support lib
API 16 is the lowest version of Android that is officially supported by
Xamarin.Android. In most cases, this restriction is only presented as a
build warning. Recent changes to use Google's [manifest merger tool][0]
have however surfaced a new error in projects which use the support
libraries and specify a minSdkVersion lower than API 14. These projects
should be updated to use a minSdkVersion of 16 to avoid both this new
error, and the existing warning about using an unsupported minimum SDK
version.

[0]: https://github.com/xamarin/xamarin-android/commit/2c6f5cd16fb5087731107c1defa13ebb56cf63df